Fixed long existing ChipSelect with reverse polarity bug.
The problem was that the chip-polarity settings are shared with the normal config of the current transaction.
but the current transaction does not know anything about the other devices,
so it resets the polarity to 0 (active on CS low), which essentially activates all devices that are active on high.

So this patch pre-calculates the relevant registers when setting up a device.
and then on transfer these values are used and filled in.

There is one thing, that might be of concern to someone:
the default "Active on Low" is applied until the first transfer happens.
but before the first request, no clock beats should show up on the bus anyway...

Signed-off-by: Martin Sperl <kernel@martin.sperl.org>