ARM: bcm2708: create a copy of bcm2835-i2s for compatiblity
To allow both kms and non kms device tree set-ups to work
with i2s-audio we need to separate the driver for the (soon)
legacy non-kms setups.

To ease the migration we keep both versions of the driver,
where the main driver will in the future support the clock
framework while the copy will continue to work with the
self-made clockmanager (which does not work when using
kms, as this requires the clock framework to be enabled).

The legacy driver now responds to the compatiblity string
"brcm,bcm2835-i2s-noclock", which is at this moment made the default
in the device-tree. the vc4-kms-v3d overlay will change the
compatiblity.

When the clock framework is enabled by default then these changes
can easily get reverted and the legacy(noclock) driver removed.

Signed-off-by: Martin Sperl <kernel@martin.sperl.org>