clk: bcm2835: mark enabled pll_dividers as critical
The bcm2835 firmware enables several clocks and plls before
booting the linux kernel.

These plls should never get disabled as it may result in a
stopped system clock and more.

So during probing we check if the pll_divider is enabled
and if it is then mark that pll_divider as critical.
As a consequence this will also enable the corresponding parent pll.

Signed-off-by: Martin Sperl <kernel@martin.sperl.org>