clk: bcm2835: mark enabled clocks with CLK_ENABLE_HAND_OFF
The bcm2835 firmware enables several clocks and plls before
booting the linux kernel.

These plls should never get disabled as it may result in a
stopped system clock and more.

So during probing we check if the clock is enabled
and if it is then mark that clock with CLK_ENABLE_HAND_OFF.
As a consequence this will also enable the corresponding
parent plls and pll-divs.

Signed-off-by: Martin Sperl <kernel@martin.sperl.org>

--
Note that this requires the following patches applied:
* clk: introduce CLK_ENABLE_HAND_OFF flag
* clk: per-user clk prepare & enable ref counts