clk: bcm2835: mark enabled clocks as critical
The bcm2835 firmware enables several clocks and plls before
booting the linux kernel.

These plls should never get disabled as it may result in a
stopped system clock and more.

So during probing we check if the clock is enabled
and if it is then mark that clock as critical.
As a consequence this will also enable the corresponding
parent plls and pll-divs.

Ideally this should use CLK_ENABLE_HAND_OFF but this is still
not merged, so until that happens we have to use
CLK_IS_CRITICAL.

Signed-off-by: Martin Sperl <kernel@martin.sperl.org>