can: mcp25xxfd: optimize SPI reads of FIFOs in can2.0 mode
To reduce the amount of spi_messages send this patch optimizes the
read of RX fifos so that multiple Fifos are read together.

This only works well in CAN2.0 mode where the overhead of prefetching
8 bytes (even if the actual length is 0) is negligable compared
to the SPI-framework overhead.

Statistics in debugfs show for 1000000 DLC=0 can messages received:
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_rx/rx_bulk_reads <==
355411
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_rx/rx_bulk_reads_1 <==
92778
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_rx/rx_bulk_reads_2 <==
7133
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_rx/rx_bulk_reads_3 <==
130390
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_rx/rx_bulk_reads_4 <==
123807
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_rx/rx_bulk_reads_5 <==
1266
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_rx/rx_bulk_reads_6 <==
34
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_rx/rx_bulk_reads_7+ <==
1
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_rx/rx_reads <==
999999
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_stats/int_system_error <==
1
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_stats/int_system_error_rx <==
1
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_stats/int_system_error_tx <==
0

There was one SERR error where we lost one can frame during the handling
of the error condition.

So to read all of those 999999 RX fifos we formerly have scheduled
999999 spi_messages.

With this patch we have only scheduled 355411 spi_messages, so
only 35.5% of the former value.

This also means we have not been transferring 1289176 (=2*(999999-355411))
unnecessary command bytes over the SPI bus.

As we have been issuing less SPI transfers reading SRAM this
also has reduced the likleyhood of SERR happening - it does not
solve the complete issue as seen above.

Signed-off-by: Martin Sperl <kernel@martin.sperl.org>

---
Changelog:
  V5->V6: new with this patchset version

Signed-off-by: Martin Sperl <kernel@martin.sperl.org>