can: mcp25xxfd: optimize tef reads reading multiple TEFs in one go
To reduce the amount of spi_messages send this patch optimizes the
read TEF so that multiple TEFs are read together.

Statistics in debugfs show for 10000 DLC=0 can messages sent:
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_tx/tef_conservative_reads
0
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_tx/tef_opportunistic_reads
44691
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_tx/tef_opportunistic_reads_1
1862
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_tx/tef_opportunistic_reads_2
30349
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_tx/tef_opportunistic_reads_3
12480
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_tx/tef_opportunistic_reads_4
0
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_tx/tef_opportunistic_reads_5
0
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_tx/tef_opportunistic_reads_6
0
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_tx/tef_opportunistic_reads_7+
0
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_tx/tef_reads
100000
==> /sys/kernel/debug/mcp25xxfd-spi0.0/can_fifo_tx/tef_read_splits
0

So to read all of those 100000 TEF frames to read we formerly have
scheduled 100000 spi_messages.
While with this patch we have only scheduled 44691 spi_messages, so
only 44.6% of the former value.
This also means we have not been transferring 110618 (=2*30349+2*2*12480)
unnecessary command bytes over the SPI bus.

Signed-off-by: Martin Sperl <kernel@martin.sperl.org>

---

Changelog:
  V5->V6: new with this patchset version